+++ components/src/index.ts
@@ -4,7 +4,9 @@
  This code is licensed under the BSD-3 license found in the LICENSE file in the root directory of this source tree and at https://opensource.org/licenses/BSD-3-Clause.
  **/
 declare global {
+    // eslint-disable-next-line @typescript-eslint/no-namespace
     namespace ReactNativePaper {
+        // eslint-disable-next-line @typescript-eslint/consistent-type-definitions
         interface ThemeColors {
             primaryBase: string;
             primaryPalette: {
@@ -43,6 +45,7 @@ declare global {
                 disabledBackground: string;
             };
         }
+        // eslint-disable-next-line @typescript-eslint/consistent-type-definitions
         interface ThemeFonts {
             bold: ThemeFont;
         }